tips_bot/requirements.txt
requests
python-dotenv
psycopg2-binary
nba_api
pandas
google-generativeai